requirements.txt
Django
psycopg2-binary
python-dotenv
djangorestframework
pillow
django-cleanup
django-rest-swagger
drf-yasg